+++ CHANGELOG.md
@@ -25,6 +25,8 @@ This is the first release to use a changelog.
 
 ### Changed
 
+- Read geometry object `meshColor` `meshScale` when updating robot scene ([#50](https://github.com/Simple-Robotics/candlewick/pull/50))
+- Set minimum version of ffmpeg to 7.x ([#50](https://github.com/Simple-Robotics/candlewick/pull/50))
 - Change signature of `loadCoalPrimitive()` to take `coal::ShapeBase`
 - Change signature of `MeshData::getAttribute()` template member function to use strided view